Linux Compatability: use 'spawn' to start new processes
On Linux, new processes start via 'fork' by default. We need it to be 'spawn' (the Windows default) in order to prevent duplicate objects in the SimModel registry.

So we get a 'spawn' context and pass it to the multiprocessing pool on initialization.

Also created __nRuns attribute in initializer (caught  by pylint)