If a message handler returns None , assume it is handled
If a message handler returns None (or returns no value at all, which translates to None), assume it is handled - i.e., don't put the message in the queue.

This is designed to ensure that client code that forgets to return a boolean doesn't inadvertently add messages to the queue.

The alternative is to raise with a return value of None, but that seems a bit extreme at this point.