simulation.py - SimulationResult - add CIs, modify
Added 90 and 95% confidence intervals as range options for summary report. We use T distribution estimate around the mean for the dataset mean value, the Conover non-parametric interval for the median for the other statistics.

For summary statistics other than the sample mean, changed to point estimate/centrality estimate to median rather than mean.